+++ examples/cc_gwlb/outputs.tf
@@ -15,18 +15,9 @@ ${join("\n", module.cc_vm.id)}
 All CC Management IPs:
 ${join("\n", module.cc_vm.private_ip)}
 
-All CC Service IPs:
+All CC Forwarding Service IPs:
 Service Interface Device Index 1:
 ${join("\n", module.cc_vm.cc_service_private_ip)} 
-
-Service Interface Device Index 2:
-${join("\n", module.cc_vm.cc_med_lrg_service_1_private_ip)} 
-
-Service Interface Device Index 3:
-${join("\n", module.cc_vm.cc_med_lrg_service_2_private_ip)} 
-
-Service Interface Device Index 4:
-${join("\n", module.cc_vm.cc_lrg_service_3_private_ip)}
 
 All CC Primary Service ENIs:
 ${join("\n", module.cc_vm.service_eni_1)}